* commit 'ccedfac1d04b7f9c4df4a144f80283f7af69a9d6':
  Fix a bunch of javac -Xlint warnings in our code.

+++ luni/src/main/java/java/util/zip/InflaterInputStream.java
@@ -199,7 +199,7 @@ public class InflaterInputStream extends FilterInputStream {
             synchronized (is.mSharedRaf) {
                 long len = is.mLength - is.mOffset;
                 if (len > nativeEndBufSize) len = nativeEndBufSize;
-                int cnt = inf.setFileInput(is.mSharedRaf.getFD(), is.mOffset, (int)nativeEndBufSize);
+                int cnt = inf.setFileInput(is.mSharedRaf.getFD(), is.mOffset, nativeEndBufSize);
                 is.skip(cnt);
             }
         } else {
